[#4626] moving schema validation to model program aggregation class
