+++ hs_file_types/models/model_program.py
@@ -2,6 +2,7 @@ import glob
 import json
 import os
 
+import jsonschema
 from urllib.parse import urlparse
 from django.conf import settings
 from django.contrib.postgres.fields import ArrayField
@@ -529,6 +530,134 @@ class ModelProgramLogicalFile(AbstractModelLogicalFile):
         This must agree between Composite Types and native types).
         """
         return "Model Program"
+
+    @staticmethod
+    def validate_meta_schema(meta_schema):
+        """A helper to validate metadata schema for model program aggregation
+        :param  meta_schema: metadata schema as a string - which needs to be validated
+        :returns a dict of metadata schema and a list of validation errors
+        """
+        validation_errors = []
+        json_schema = dict()
+        is_schema_valid = True
+        try:
+            json_schema = json.loads(meta_schema)
+        except ValueError:
+            validation_errors.append("Schema is not valid JSON")
+            return json_schema, validation_errors
+
+        if json_schema:
+            schema_version = json_schema.get("$schema", "")
+            if not schema_version:
+                is_schema_valid = False
+                err_message = "Not a valid JSON schema. {}"
+                if "$schema" not in json_schema:
+                    err_message = err_message.format("Key '$schema' is missing")
+                else:
+                    err_message = err_message.format("Key '$schema' is missing a value for schema version")
+                validation_errors.append(err_message)
+            else:
+                if "/draft-04/" not in schema_version:
+                    is_schema_valid = False
+                    err_message = "Not a valid JSON schema. Schema version is invalid. Supported valid version(s): " \
+                                  "draft-04"
+                    validation_errors.append(err_message)
+
+            if 'properties' not in json_schema:
+                is_schema_valid = False
+                validation_errors.append("Not a valid metadata schema. Attribute 'properties' is missing")
+
+            if is_schema_valid:
+                try:
+                    jsonschema.Draft4Validator.check_schema(json_schema)
+                except jsonschema.SchemaError as ex:
+                    is_schema_valid = False
+                    schema_err_msg = f"{ex.message}. Schema invalid field path:{str(list(ex.path))}"
+                    schema_err_msg = f"Not a valid JSON schema. Error:{schema_err_msg}"
+                    validation_errors.append(schema_err_msg)
+
+        if is_schema_valid:
+            # custom validation - hydroshare requirements
+            # this custom validation requiring additional attributes are needed for making the json-editor form
+            # generation at the front-end to work
+            if 'additionalProperties' not in json_schema:
+                is_schema_valid = False
+                validation_errors.append("Not a valid metadata schema. Attribute 'additionalProperties' is missing")
+
+            elif json_schema['additionalProperties']:
+                is_schema_valid = False
+                validation_errors.append("Not a valid metadata schema. Attribute 'additionalProperties' "
+                                         "should bet set to 'false'")
+
+            def validate_schema(schema_dict):
+                for k, v in schema_dict.items():
+                    # key must not have whitespaces - required for xml encoding of metadata
+                    if k != k.strip():
+                        msg = f"Not a valid metadata schema. Attribute '{k}' has leading or trailing whitespaces"
+                        validation_errors.append(msg)
+                    # key must consists of alphanumeric characters only - required for xml encoding of metadata
+                    if not k.isalnum():
+                        msg = f"Not a valid metadata schema. Attribute '{k}' has non-alphanumeric characters"
+                        validation_errors.append(msg)
+
+                    # key must start with a alphabet character - required for xml encoding of metadata
+                    if not k[0].isalpha():
+                        msg = f"Not a valid metadata schema. Attribute '{k}' starts with a non-alphabet character"
+                        validation_errors.append(msg)
+
+                    if isinstance(v, dict):
+                        if k not in ('properties', 'items'):
+                            # we need a title to use as label for the form field
+                            if 'title' not in v:
+                                msg = f"Not a valid metadata schema. Attribute 'title' is missing for {k}"
+                                validation_errors.append(msg)
+                            elif len(v['title'].strip()) == 0:
+                                msg = f"Not a valid metadata schema. Attribute 'title' has no value for {k}"
+                                validation_errors.append(msg)
+
+                            if v['type'] == 'array':
+                                # we need format attribute set to 'table' in order for the jsoneditor to allow
+                                # editing array type field
+                                if 'format' not in v:
+                                    msg = f"Not a valid metadata schema. Attribute 'format' is missing for {k}"
+                                    validation_errors.append(msg)
+
+                                elif v['format'] != 'table':
+                                    msg = f"Not a valid metadata schema. Attribute 'format' should be set " \
+                                          f"to table for {k}"
+                                    validation_errors.append(msg)
+
+                        if 'type' in v and v['type'] == 'object':
+                            # we requiring "additionalProperties": false so that we don't allow user to add new
+                            # form fields using the json-editor form
+                            if 'additionalProperties' not in v:
+                                msg = "Not a valid metadata schema. Attribute 'additionalProperties' is " \
+                                      f"missing for {k}"
+                                validation_errors.append(msg)
+                            elif v['additionalProperties']:
+                                msg = "Not a valid metadata schema. Attribute 'additionalProperties' must " \
+                                      f"be set to false for {k}"
+                                validation_errors.append(msg)
+
+                        # check for nested objects - we are not allowing nested objects to keep the form
+                        # generated from the schema by json-editor to not get complicated
+                        nested_object_found = False
+                        if 'type' in v and v['type'] == 'object':
+                            # parent type is object - check child type is not object
+                            for k_child, v_child in v.items():
+                                if isinstance(v_child, dict):
+                                    if 'type' in v_child and v_child['type'] == 'object':
+                                        msg = "Not a valid metadata schema. Nested object types are not allowed. " \
+                                              f"Attribute '{k_child}' contains nested object types"
+                                        validation_errors.append(msg)
+                                        nested_object_found = True
+                        if not nested_object_found:
+                            validate_schema(v)
+
+            if is_schema_valid:
+                validate_schema(json_schema['properties'])
+
+        return json_schema, validation_errors
 
     def copy_mp_file_types(self, tgt_logical_file):
         """helper function to support creating copy or new version of a resource
